Update various substitutions files
This commit moves around all of the information for the various substitutions files.  The new structure simplifies the substitutions for the 'location' of the device.  The offset for reading information from each channel of the controller are removed from the substitutions file.  The offsets for writing to various values are now controlled by the substitutions file.